change type of var choice from long to int

+++ testsingle.c
@@ -5,7 +5,7 @@
 int main()
 {
     int ret, fd=0;
-    long choice=0;
+    int choice=0;
     size_t len=1;
     char str[256];
     while (choice != 5) {
